requirements.txt
numpy
tweepy
scikit-learn
python-dotenv
fastapi[all]
